backend-ORM-parking/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
jinja2
python-multipart